Route through WalletConnectMainFragment
