+++ app/src/main/AndroidManifest.xml
@@ -58,9 +58,6 @@
         <activity
             android:name=".modules.keystore.KeyStoreActivity"
             android:screenOrientation="portrait" />
-        <activity
-            android:name=".modules.walletconnect.WalletConnectActivity"
-            android:screenOrientation="portrait" />
 
         <service
             android:name=".core.fcm.MyFirebaseMessagingService"
