+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/walletconnect/WalletConnectErrorFragment.kt
@@ -4,6 +4,7 @@ import android.os.Bundle
 import android.view.View
 import androidx.core.os.bundleOf
 import androidx.fragment.app.Fragment
+import androidx.navigation.fragment.findNavController
 import io.horizontalsystems.bankwallet.R
 import io.horizontalsystems.bankwallet.core.setOnSingleClickListener
 import kotlinx.android.synthetic.main.fragment_wallet_connect_error.*
@@ -16,18 +17,12 @@ class WalletConnectErrorFragment : Fragment(R.layout.fragment_wallet_connect_err
         message.text = arguments?.getString(MESSAGE_KEY)
 
         cancelButton.setOnSingleClickListener {
-            requireActivity().finish()
+            findNavController().popBackStack()
         }
     }
 
     companion object {
         const val MESSAGE_KEY = "MESSAGE_KEY"
-
-        fun newInstance(message: String): WalletConnectErrorFragment {
-            return WalletConnectErrorFragment().apply {
-                arguments = bundleOf(MESSAGE_KEY to message)
-            }
-        }
     }
 
 }
