+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/walletconnect/WalletConnectModule.kt
@@ -17,8 +17,4 @@ object WalletConnectModule {
         }
     }
 
-    fun start(activity: FragmentActivity) {
-        activity.startActivity(Intent(activity, WalletConnectActivity::class.java))
-    }
-
 }
